requirements.txt
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.2.0
exceptiongroup~=1.2.2
sniffio~=1.3.1
fastapi~=0.115.8
starlette~=0.44.0
pydantic~=2.10.6
anyio~=4.5.2
annotated-types~=0.7.0
pyspark~=3.5.4
kafka-python~=2.0.3

h11~=0.14.0
py4j~=0.10.9.7
pytz~=2025.1
numpy~=1.24.4
charset-normalizer~=3.4.1
pandas~=2.0.3
python-dateutil~=2.9.0.post0
certifi~=2025.1.31
requests~=2.32.3
pyspark-stubs~=3.0.0.post3
urllib3~=2.2.3
click~=8.1.8
uvicorn~=0.33.0
six~=1.17.0
aiohttp~=3.10.11
geoip2~=4.8.1